+++ chunker/src/model/function/details/function_post_np.cpp
@@ -5,8 +5,13 @@ namespace xforce { namespace nlu { namespace chunker {
 void FunctionPostNp::Process(
         const milkie::StorageItem &storageItem,
         basic::NluContext &nluContext) {
-
+  nluContext.GetChunkSeps().Add(
+          std::make_shared<basic::ChunkSep>(
+                  storageItem.GetOffset() + storageItem.GetContent().length()));
 }
 
+bool FunctionPostNp::ProcessModifier_() {
+
+}
 
 }}}
